+++ src/frontend/packages/lib_classroom/src/components/ClassroomWidgetProvider/widgets/Recordings/index.tsx
@@ -53,6 +53,7 @@ export const Recordings = () => {
             key={recording.id}
             recording={recording}
             classroomTitle={classroom.title}
+            conversionEnabled={classroom.vod_conversion_enabled}
           />
         )}
       </ItemList>